Partial egg update

+++ server_src/monitor3.c
@@ -6,7 +6,7 @@
 /*   By: dslogrov <dslogrove@gmail.com>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/26 15:33:13 by dslogrov          #+#    #+#             */
-/*   Updated: 2019/12/04 18:01:35 by dslogrov         ###   ########.fr       */
+/*   Updated: 2019/12/09 11:10:01 by dslogrov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,12 +30,12 @@ void	monitor_pbc(t_state *s, t_player *p, char *message)
 void	monitor_pic(t_state *s, t_player *player1)
 {
 	char		buff[STRBUFF_SIZE];
-	size_t		i;
+	int			i;
 	t_player	*player;
 
 	snprintf(buff, STRBUFF_SIZE, "pic %u %u %u", player1->x, player1->y, player1->level);
 	send_all_monitors(s, buff);
-	i = -1UL;
+	i = -1;
 	while(++i <= s->max_fd)
 	{
 		if (s->clients[i].type != PLAYER)
